Add files via upload
This is a pdf document that contains the problem requirements, the conceptual model in UML for this project, the descriptions of the functionalities that I have selected to be used as an in-memory key-value storage, the descriptions of the Redis data structures that I am going to use to implement the functionalities that I have described in the previous point, and the descriptions of all the Redis commands that I would like to use to interact with my specific Redis data structures. I have described the commands for all my use cases and have included all of the CRUD operations for each of the in-memory key-value storages that I would be using.